+++ scripts/domainwall_pinning/const_a/const_a_dwp.py
@@ -44,7 +44,7 @@ H_analytic = H(soft_Aex, soft_ms, soft_K_uni, hard_Aex, hard_ms, hard_K_uni)*Con
 print("H_analytic=", H_analytic, " [T]")
 
 # Creating objects
-m = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f64)
+m = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f32)
 m[:nx/2, :, :, 0] =  1.0
 m[:nx/2, :, :, 1] =  0.3
 m[nx/2:, :, :, 0] = -1.0
@@ -53,21 +53,21 @@ m[nx/2:, :, :, 1] =  0.3
 mesh = Mesh(nx, ny, nz, x/nx, y/ny, z/nz)
 
 # Setting A values as field
-A_field = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f64)
+A_field = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f32)
 # soft
 A_field[:nx/2, :, :, :] = soft_Aex
 # hard
 A_field[nx/2:, :, :, :] = hard_Aex
 
 # Setting Ms values as field
-Ms_field = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f64)
+Ms_field = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f32)
 # soft
 Ms_field[:nx/2, :, :, :] = soft_ms
 # hard
 Ms_field[nx/2:, :, :, :] = hard_ms
 
 # Setting Ku1 values as field
-Ku1_field = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f64)
+Ku1_field = af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f32)
 # soft
 Ku1_field[:nx/2, :, :, :] = soft_K_uni
 # hard
@@ -77,7 +77,7 @@ state = State(mesh, Ms_field, m)
 state.normalize()
 state.write_vti(sys.argv[1] + "minit")
 fields = [
-    ExternalField(af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f64)),
+    ExternalField(af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f32)),
     ExchangeField(A_field),
     UniaxialAnisotropyField(Ku1_field, Ku1_axis=[1., 0., 0.]),
     #DemagField(mesh),
