+++ scripts/elliptical/ellipse_fixed_layer.py
@@ -44,8 +44,8 @@ if fixed_is_elliptical == True:
   fixed_layer, n_cells_2  = Util.disk(nx, ny, 1, axis=[0, 0, 1])
   disk[:, :, 0 , :] = fixed_layer
 else:
-  fixed_layer = af.constant(0., nx, ny, 1, 3, dtype=af.Dtype.f64)
-  fixed_layer[:, :, 0 , 2] = af.constant(1., nx, ny, 1, 1, dtype=af.Dtype.f64)
+  fixed_layer = af.constant(0., nx, ny, 1, 3, dtype=af.Dtype.f32)
+  fixed_layer[:, :, 0 , 2] = af.constant(1., nx, ny, 1, 1, dtype=af.Dtype.f32)
   disk[:, :, 0 , :] = fixed_layer
 
 
@@ -59,7 +59,7 @@ fields = [
     DemagField(mesh, material, verbose = True),
     ExchangeField(mesh, material),
     #UniaxialAnisotropyField(mesh, material),
-    #ExternalField(af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f64))
+    #ExternalField(af.constant(0.0, nx, ny, nz, 3, dtype=af.Dtype.f32))
 ]
 print ("Initialized interaction terms in ", time.time() - start, "[s]")
 Llg = LLGIntegrator(terms = fields)
